board/via/imx6_vab820: update to latest BSP release
Changes:
* Update Kernel and U-Boot to the latest 2.0.6 BSP release from the
official sources.
* Add genimage config to create bootable SD card image

---
I've added the genimage method because it is part very handy, and
part seem to have shown up for a couple of other boards in patches,
though not in the repo. Let me know if that's okay to have as default,
or if I should roll back that part of the changeset and just have
manual boot image creation steps like before.